requirements.txt
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
faiss-cpu==1.9.0.post1
gitpython==3.1.44
streamlit==1.41.1
python-dotenv==1.0.1
tiktoken==0.8.0